refactor: move cleanup work to more general location
