+++ packages/core/src/shared/index.ts
@@ -63,6 +63,7 @@ export { TabTypeDataMap } from '../amazonq/webview/ui/tabs/constants'
 export * from './lsp/manifestResolver'
 export * from './lsp/lspResolver'
 export * from './lsp/types'
+export * from './lsp/utils/cleanup'
 export { default as request } from './request'
 export * from './lsp/utils/platform'
 export * as processUtils from './utilities/processUtils'
